updated files css and widget
