+++ etc/widget.xml
@@ -1,11 +1,17 @@
 <?xml version="1.0" ?>
 <widgets xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Widget:etc/widget.xsd">
 	<widget class="Lof\ProductTags\Block\Widget\LofproductTags" id="lof_producttags_lofproducttags">
-		<label>lofproductTags</label>
-		<description>lofproductTags</description>
+		<label>Display Product Tags</label>
+		<description>Display Product Tags Widget</description>
 		<parameters>
-			<parameter name="title" sort_order="5" visible="true" xsi:type="text">
-				<label>title</label>
+			<parameter name="title" visible="true" xsi:type="text">
+				<label>Widget Title</label>
+			</parameter>
+			<parameter name="number_tags" visible="true" xsi:type="text">
+				<label>Limit tags to show</label>
+			</parameter>
+			<parameter name="show_number_products" xsi:type="select"  visible="true" source_model="Magento\Config\Model\Config\Source\Yesno">
+				<label translate="true">Show Number Of Products</label>
 			</parameter>
 		</parameters>
 	</widget>
