+++ view/frontend/templates/listtags.phtml
@@ -5,5 +5,4 @@
 ?>
 <div>
 	<?= $block->getListTags() ?>
-	<?= __('Hello Lof_ProductTags::listtags.phtml') ?>
 </div>
